--- a/src/sparsify/auto/scripts/main.py
+++ b/src/sparsify/auto/scripts/main.py
@@ -26,9 +26,10 @@ from sparsify.auto.utils import (
 from sparsify.schemas import APIArgs
 from sparsify.schemas.auto_api import SparsificationTrainingConfig
 from tensorboard.program import TensorBoard
+from tensorboard.util import tb_logging
 
 
-_LOGGER = logging.getLogger("auto_banner")
+_LOGGER = tb_logging.get_logger()
 
 
 def main(api_args: APIArgs):
@@ -41,6 +42,11 @@ def main(api_args: APIArgs):
         deploy_directory,
     ) = create_save_directory(api_args)
 
+    logging_level = logging.WARNING
+    if "logging_level" in api_args.kwargs:
+        logging_level = api_args.kwargs["logging_level"]
+
+    _LOGGER.setLevel(logging_level)
     # Launch tensorboard server
     tensorboard_server = TensorBoard()
     tensorboard_server.configure(argv=[None, "--logdir", log_directory])